Expose access to sub objects/classes
The functions which process glade files now return a dictionary mapping
class names to classes. Instances of these classes have a method
``get_object``, which takes the name of a class (as defined in the glade
file), and returns the instance of that class that was created at the
same time as the instance the method was called on.